move api/token.go -> api/auth/token.go
