+++ api/auth/token.go
@@ -2,7 +2,7 @@
 //
 // Use of this source code is governed by the LICENSE file in this repository.
 
-package api
+package auth
 
 import (
 	"fmt"
